refactor: split storage package into domain files
Move *Storage methods out of storage.go into user/thumbnail/tag/title/identity/missing/library_cache files without API changes.
